test: Consolidate modules for testing of grid.
